+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 2.2.5 / Unreleased
+## 2.3.0 / Unreleased
+
+- #80: Drop support for Python 3.3 (end-of-life september 2017)
+- #86: Give custom PropertyManager implementations access to the environ
+- #87: Don't assume sys.stdout.encoding is not None
+
 
 ## 2.2.4 / 2017-08-11